refactor: :recycle: refactor file structure for apps and componentsThis commit renames several files and moves them to a more organizeddirectory structure under `src/apps` and `src/components`. Thisimproves readability and maintainability of the codebase
