+++ src/apps/ModalApp.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState } from 'react'
 import ModifySchedule, { IScheduleValue } from '@/components/ModifySchedule'
-import TodayTaskModal from './components/TodayTaskModal'
+import TodayTaskModal from '../components/TodayTaskModal'
 import { useAtom } from 'jotai'
-import { projectSchedulesAtom } from './model/schedule'
-import { getInternalEvents, IEvent } from './util/events'
-import { fullEventsAtom, journalEventsAtom, projectEventsAtom } from './model/events'
-import PomodoroModal from './components/PomodoroModal'
+import { projectSchedulesAtom } from '../model/schedule'
+import { getInternalEvents, IEvent } from '../util/events'
+import { fullEventsAtom, journalEventsAtom, projectEventsAtom } from '../model/events'
+import PomodoroModal from '../components/PomodoroModal'
 import { ConfigProvider } from 'antd'
-import { ANTD_THEME_CONFIG } from './util/constants'
-import useTheme from './hooks/useTheme'
-import AddDailyLogModal from './components/AddDailyLogModal'
+import { ANTD_THEME_CONFIG } from '../util/constants'
+import useTheme from '../hooks/useTheme'
+import AddDailyLogModal from '../components/AddDailyLogModal'
 
 type IEditSchedule = {
   type: 'editSchedule'
