+++ src/apps/PomodoroApp.tsx
@@ -3,17 +3,17 @@ import { PomodoroConfig, usePomodoro } from '@haydenull/use-pomodoro'
 import { RiExternalLinkLine } from 'react-icons/ri'
 import { MdOutlineRunningWithErrors } from 'react-icons/md'
 import { AiOutlinePoweroff, AiOutlineQuestionCircle } from 'react-icons/ai'
-import { transformBlockToEvent } from './helper/transform'
-import { getInitialSettings } from './util/baseInfo'
-import { IEvent } from './util/events'
+import { transformBlockToEvent } from '../helper/transform'
+import { getInitialSettings } from '../util/baseInfo'
+import { IEvent } from '../util/events'
 import { genToolbarPomodoro, secondsToTime, updatePomodoroInfo } from '@/helper/pomodoro'
-import { getPageData, navToBlock } from './util/logseq'
+import { getPageData, navToBlock } from '../util/logseq'
 import dayjs from 'dayjs'
-import { notification } from './util/util'
-import InerruptionModal from './components/InerruptionModal'
-import useTheme from './hooks/useTheme'
+import { notification } from '../util/util'
+import InerruptionModal from '../components/InerruptionModal'
+import useTheme from '../hooks/useTheme'
 import { ConfigProvider, Tooltip } from 'antd'
-import { ANTD_THEME_CONFIG } from './util/constants'
+import { ANTD_THEME_CONFIG } from '../util/constants'
 
 const OperationButton = ({ label, onClick, classNames = '', tooltip }: {label: string; onClick: () => void; classNames?: string; tooltip?: string}) => (
   <button
