+++ src/main.tsx
@@ -1,49 +1,28 @@
 import '@logseq/libs'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import dayjs from 'dayjs'
-import weekday from 'dayjs/plugin/weekday'
-import updateLocale from 'dayjs/plugin/updateLocale'
-import isSameOrBefore from 'dayjs/plugin/isSameOrBefore'
-import isSameOrAfter from 'dayjs/plugin/isSameOrAfter'
-import localeData from 'dayjs/plugin/localeData'
-import difference from 'lodash/difference'
-import findKey from 'lodash/findKey'
-import isBetween from 'dayjs/plugin/isBetween'
-import utc from 'dayjs/plugin/utc'
-import quarterOfYear from 'dayjs/plugin/quarterOfYear'
 import * as echarts from 'echarts/core'
 import { GridComponent, ToolboxComponent, TooltipComponent, LegendComponent} from 'echarts/components'
 import { LineChart, GaugeChart, BarChart, TreemapChart, PieChart } from 'echarts/charts'
 import { UniversalTransition } from 'echarts/features'
 import { CanvasRenderer } from 'echarts/renderers'
 import 'antd/dist/reset.css'
-import { getInitialSettings, initializeSettings } from './util/baseInfo'
+import { getInitialSettings, initializeSettings } from '@/util/baseInfo'
 import App from './App'
 import 'tui-calendar/dist/tui-calendar.css'
 import './style/index.less'
-import { listenEsc, log, managePluginTheme, setPluginTheme, toggleAppTransparent } from './util/util'
-import { genToolbarPomodoro, togglePomodoro } from '@/helper/pomodoro'
-import ModalApp, { IModalAppProps } from './ModalApp'
-import TaskListApp from './TaskListApp'
-import { genDBTaskChangeCallback, getBlockUuidFromEventPath } from './util/logseq'
-import { LOGSEQ_PROVIDE_COMMON_STYLE } from './constants/style'
-import { transformBlockToEvent } from './helper/transform'
-import PomodoroApp from './PomodoroApp'
-import { pullTask, getTodoistInstance, updateTask, closeTask, getTask, reopenTask, createTask, updateBlock, PRIORITY_MAP, transformEventToTodoistEvent } from './helper/todoist'
-import { AddTaskArgs, TodoistRequestError, UpdateTaskArgs } from '@doist/todoist-api-typescript'
-import { DEFAULT_PROJECT } from './util/constants'
-import initializeDayjs from './register/dayjs'
-
-dayjs.extend(weekday)
-dayjs.extend(isSameOrBefore)
-dayjs.extend(isSameOrAfter)
-dayjs.extend(localeData)
-dayjs.extend(difference)
-dayjs.extend(isBetween)
-dayjs.extend(updateLocale)
-dayjs.extend(utc)
-dayjs.extend(quarterOfYear)
+import { listenEsc, log, managePluginTheme, setPluginTheme, toggleAppTransparent } from '@/util/util'
+import { togglePomodoro } from '@/helper/pomodoro'
+import ModalApp, { IModalAppProps } from '@/apps/ModalApp'
+import { LOGSEQ_PROVIDE_COMMON_STYLE } from '@/constants/style'
+import { pullTask } from '@/helper/todoist'
+import initializeDayjs from '@/register/dayjs'
+import initializeTodoist from '@/register/todoist'
+import initializeSidebar from '@/register/sidebar'
+import initializePomodoro, { renderPomodoroApp } from '@/register/pomodoro'
+import initializeClickLogseqDomListener from '@/register/logseq/clickLogseqDom'
+import initializePageMenuItem from '@/register/pageMenuItem'
+import initializeEditScheduleModal from '@/register/editScheduleModal'
 
 echarts.use([GridComponent, LineChart, BarChart, GaugeChart, TreemapChart, PieChart, CanvasRenderer, UniversalTransition, ToolboxComponent, TooltipComponent, LegendComponent])
 
@@ -56,36 +35,18 @@ if (isDevelopment) {
 } else {
   log('=== logseq-plugin-agenda loaded ===')
   logseq.ready(() => {
-
     initializeSettings()
-
-    const { weekStartDay, todoist } = getInitialSettings()
-
-    // dayjs.updateLocale('en', {
-    //   weekStart: weekStartDay,
-    // })
-    initializeDayjs(weekStartDay)
-
-    managePluginTheme()
-
-    Notification.requestPermission()
-
     logseq.App.getUserConfigs().then(configs => {
       window.logseqAppUserConfigs = configs
     })
+    // fix: https://github.com/haydenull/logseq-plugin-agenda/issues/87
+    logseq.setMainUIInlineStyle({ zIndex: 1000 })
+    logseq.provideStyle(LOGSEQ_PROVIDE_COMMON_STYLE)
+    Notification.requestPermission()
 
-    listenEsc(() => logseq.hideMainUI())
+    const { weekStartDay, todoist } = getInitialSettings()
 
-    logseq.App.onThemeModeChanged(({ mode }) => {
-      setPluginTheme(mode)
-    })
-
-    logseq.on('ui:visible:changed', (e) => {
-      if (!e.visible && window.currentApp !== 'pomodoro') {
-        ReactDOM.unmountComponentAtNode(document.getElementById('root') as Element)
-      }
-    })
-
+    // ===== logseq plugin model start =====
     logseq.provideModel({
       show() {
         renderApp()
@@ -104,85 +65,18 @@ if (isDevelopment) {
         pullTask()
       },
     })
-
+    // ===== logseq plugin model end =====
+    // ========== show or hide app start =========
     logseq.App.registerUIItem('toolbar', {
       key: 'logseq-plugin-agenda',
       template: '<a data-on-click="show" class="button"><i class="ti ti-comet"></i></a>',
     })
-
-    // ========== todoist =========
-    if (todoist?.token) {
-      logseq.App.registerUIItem('toolbar', {
-        key: 'plugin-agenda-todoist',
-        template: '<a data-on-click="pullTodoistTasks" class="button"><i class="ti ti-chevrons-down"></i></a>',
-      })
-      getTodoistInstance()
-
-      logseq.DB.onChanged(({ blocks, txData, txMeta }) => {
-        // console.log('[faiz:] === main DB.onChanged', blocks, txData, txMeta)
-        const syncToTodoist = async (uuid: string) => {
-          const block = await logseq.Editor.getBlock(uuid)
-          const event = await transformBlockToEvent(block!, getInitialSettings())
-
-          const todoistId = transformEventToTodoistEvent(event)?.todoistId
-          try {
-            const task = await getTask(todoistId)
-            const priority = findKey(PRIORITY_MAP, v => v === event.priority)
-            let params: UpdateTaskArgs = {
-              content: event.addOns.contentWithoutTime?.split('\n')?.[0],
-              description: block?.properties?.todoistDesc,
-              priority,
-            }
-            if (event.addOns.allDay === true) params.dueDate = dayjs(event.addOns.start).format('YYYY-MM-DD')
-            if (event.addOns.allDay === false) params.dueDatetime = dayjs.utc(event.addOns.start).format()
-            if (!event.rawTime) params.dueString = 'no due date'
-            if (event.addOns.status === 'done' && task?.isCompleted === false) return closeTask(todoistId)
-            if (event.addOns.status !== 'done' && task?.isCompleted === true) return reopenTask(todoistId)
-            updateTask(todoistId, params)
-          } catch (error) {
-            if ((error as TodoistRequestError).httpStatusCode === 404) {
-              return logseq.UI.showMsg(`Todoist Sync Error\nmessage: ${(error as TodoistRequestError)?.responseData}\nPlease check whether the task has been deleted or whether the todoist-id is correct`, 'error')
-            }
-          }
-        }
-        genDBTaskChangeCallback(syncToTodoist)?.({ blocks, txData, txMeta })
-      })
-
-      // @ts-ignore The requirement to return a void can be ignored
-      logseq.Editor.registerBlockContextMenuItem('Agenda: Upload to todoist', async ({ uuid }) => {
-        const settings = getInitialSettings()
-        const { todoist } = settings
-        const block = await logseq.Editor.getBlock(uuid)
-        if (!block?.marker) return logseq.UI.showMsg('This block is not a task', 'error')
-        if (block?.properties?.todoistId) return logseq.UI.showMsg('This task has already been uploaded,\nplease do not upload it again', 'error')
-        const event = await transformBlockToEvent(block!, settings)
-
-        const priority = findKey(PRIORITY_MAP, v => v === event.priority)
-
-        let params: AddTaskArgs = {
-          content: event.addOns.contentWithoutTime?.split('\n')?.[0],
-          description: block?.properties?.todoistDesc,
-          priority,
-        }
-        if (event.addOns.allDay === true) params.dueDate = dayjs(event.addOns.start).format('YYYY-MM-DD')
-        if (event.addOns.allDay === false) params.dueDatetime = dayjs.utc(event.addOns.start).format()
-        if (todoist?.project) params.projectId = todoist.project
-        if (todoist?.label) params.labels = [todoist.label]
-
-        createTask(params)
-          ?.then(async task => {
-            await updateBlock(event, task)
-            return logseq.UI.showMsg('Upload task to todoist success')
-          })
-          .catch(err => {
-            logseq.UI.showMsg('Upload task to todoist failed', 'error')
-            console.error('[faiz:] === Upload task to todoist failed', err)
-          })
-
-      })
-    }
-    // ========== todoist ==============
-
+    logseq.on('ui:visible:changed', (e) => {
+      if (!e.visible && window.currentApp !== 'pomodoro') {
+        ReactDOM.unmountComponentAtNode(document.getElementById('root') as Element)
+      }
+    })
+    listenEsc(() => logseq.hideMainUI())
     logseq.App.registerCommandPalette({
       key: 'logseq-plugin-agenda:show',
       label: 'Show Agenda',
@@ -193,56 +87,9 @@ if (isDevelopment) {
       renderApp()
       logseq.showMainUI()
     })
+    // ========== show or hide app end =========
 
-    window.unmountPomodoroApp = () => ReactDOM.unmountComponentAtNode(document.getElementById('pomodoro-root') as Element)
-    window.interruptionMap = new Map()
-
-    const editSchedule = async (e) => {
-      let block = await logseq.Editor.getBlock(e.uuid)
-      const blockRefs = await Promise.all(block!.refs?.map(ref => logseq.Editor.getPage(ref.id)))
-      block!.refs = blockRefs
-      const event = await transformBlockToEvent(block!, getInitialSettings())
-      renderModalApp({
-        type: 'editSchedule',
-        data: {
-          type: 'update',
-          initialValues: {
-            id: event.uuid,
-            start: dayjs(event.addOns.start),
-            end: dayjs(event.addOns.end),
-            isAllDay: event.addOns.allDay,
-            calendarId: event.addOns.calendarConfig?.id,
-            title: event.addOns.showTitle,
-            // keepRef: schedule.calendarId?.toLowerCase() === 'journal',
-            raw: event,
-          },
-        },
-        showKeepRef: true,
-      })
-      logseq.showMainUI()
-    }
-    logseq.App.registerPageMenuItem('Agenda: Add this page to agenda project', async ({ page }) => {
-      const pageData = await logseq.Editor.getPage(page)
-      if (!pageData) return logseq.UI.showMsg('Page not found', 'error')
-      const originalProjectList = logseq.settings?.projectList || []
-      const pageName = pageData.originalName
-      if (originalProjectList.find(project => project.id === pageName)) return logseq.UI.showMsg('This Page is already in agenda project', 'warning')
-      const newProject = {
-        ...DEFAULT_PROJECT,
-        id: pageName,
-      }
-      // hack https://github.com/logseq/logseq/issues/4447
-      logseq.updateSettings({projectList: 1})
-      // ensure subscription list is array
-      logseq.updateSettings({ ...logseq.settings, projectList: originalProjectList.concat(newProject)})
-      logseq.UI.showMsg('Successfully added')
-    })
-    logseq.Editor.registerBlockContextMenuItem('Agenda: Modify Schedule', editSchedule)
-    // @ts-ignore The requirement to return a void can be ignored
-    logseq.Editor.registerBlockContextMenuItem('Agenda: Start Pomodoro Timer', async ({ uuid }) => {
-      startPomodoro(uuid)
-    })
-    logseq.Editor.registerSlashCommand('Agenda: Modify Schedule', editSchedule)
+    // ========== today task list start =========
     logseq.Editor.registerSlashCommand("Agenda: Insert Today's Task", (e) => {
       renderModalApp({
         type: 'insertTodaySchedule',
@@ -253,99 +100,37 @@ if (isDevelopment) {
       logseq.showMainUI()
       return Promise.resolve()
     })
-    logseq.Editor.registerSlashCommand('Agenda: Insert Task List', async () => {
-      logseq.Editor.insertAtEditingCursor(`{{renderer agenda, task-list}}`)
+    // ========== today task list end =========
+
+    // ========== dayjs start =========
+    initializeDayjs(weekStartDay)
+    // ========== dayjs end =========
+
+    // ========== manage theme start =========
+    managePluginTheme()
+    logseq.App.onThemeModeChanged(({ mode }) => {
+      setPluginTheme(mode)
     })
-    logseq.App.onMacroRendererSlotted(async ({ slot, payload: { arguments: args, uuid } }) => {
-      if (args?.[0] !== 'agenda') return
-      if (args?.[1] === 'task-list') {
-        const rendered = parent.document.getElementById(slot)?.childElementCount
-        if (rendered) return
+    // ========== manage theme end ==============
 
-        const id = `agenda-task-list-${slot}`
-        logseq.provideUI({
-          key: `agenda-${slot}`,
-          slot,
-          reset: true,
-          template: `<div id="${id}"></div>`,
-          // style: {},
-        })
-        logseq.provideStyle(`${LOGSEQ_PROVIDE_COMMON_STYLE}
-          #block-content-${uuid} .lsp-hook-ui-slot {
-            width: 100%;
-          }
-          #block-content-${uuid} .lsp-hook-ui-slot > div {
-            width: 100%;
-          }
-          #block-content-${uuid} .lsp-hook-ui-slot > div > div {
-            width: 100%;
-          }
-        `)
-        setTimeout(() => {
-          ReactDOM.render(
-            <React.StrictMode>
-              <TaskListApp containerId={id} />
-            </React.StrictMode>,
-            parent.document.getElementById(id)
-          )
-        }, 0)
-      } else if (args?.[1] === 'pomodoro-timer') {
-        const id = `agenda-pomodoro-timer-${slot}`
-        logseq.provideUI({
-          key: `agenda-${slot}`,
-          slot,
-          reset: true,
-          template: `<div id="${id}"></div>`,
-          // style: {},
-        })
-      }
-    })
+    // ========== todoist =========
+    initializeTodoist(todoist)
+    // ========== todoist ==============
 
-    // fix: https://github.com/haydenull/logseq-plugin-agenda/issues/87
-    logseq.setMainUIInlineStyle({ zIndex: 1000 })
-    logseq.provideStyle(LOGSEQ_PROVIDE_COMMON_STYLE)
+    // ========== sidebar =========
+    initializeSidebar()
+    // ========== sidebar ==============
 
-    if (top) {
-      top.document.addEventListener('click', async e => {
-        const path = e.composedPath()
-        const target = path[0] as HTMLAnchorElement
-        if (target.tagName === 'A' && target.className.includes('external-link') && target.getAttribute('href')?.startsWith('#agenda')) {
-          let modalType = 'agenda'
-          if (target.getAttribute('href')?.startsWith('#agenda-pomo://')) modalType = 'pomodoro'
+    // ========== pomodoro =========
+    initializePomodoro()
+    // ========== pomodoro ==============
 
-          const uuid = getBlockUuidFromEventPath(path as unknown as HTMLElement[])
-          if (!uuid) return
-          const block = await logseq.Editor.getBlock(uuid)
-          const page = await logseq.Editor.getPage(block!.page?.id)
-          const event = await transformBlockToEvent(block!, getInitialSettings())
-          if (modalType === 'agenda') {
-            renderModalApp({
-              type: 'editSchedule',
-              data: {
-                type: 'update',
-                initialValues: {
-                  id: uuid,
-                  title: event.addOns.showTitle,
-                  calendarId: page?.originalName,
-                  keepRef: false,
-                  start: dayjs(event.addOns.start),
-                  end: dayjs(event.addOns.end),
-                  isAllDay: event.addOns.allDay,
-                  raw: event,
-                },
-              },
-            })
-          } else if (modalType === 'pomodoro') {
-            renderModalApp({
-              type: 'pomodoro',
-              data: event,
-            })
-          }
-          logseq.showMainUI()
-        }
-      })
-    }
+    // ========== listen click logseq dom =========
+    initializeClickLogseqDomListener()
+    // ========== listen click logseq dom ==============
 
+    initializePageMenuItem()
+    initializeEditScheduleModal()
   })
 }
 
@@ -378,28 +163,4 @@ export function renderModalApp(params: IModalAppProps) {
     </React.StrictMode>,
     document.getElementById('root')
   )
-}
-
-function renderPomodoroApp(uuid: string) {
-  window.currentApp = 'pomodoro'
-  togglePomodoro(true)
-  toggleAppTransparent(true)
-  ReactDOM.render(
-    <React.StrictMode>
-      <PomodoroApp uuid={uuid} />
-    </React.StrictMode>,
-    document.getElementById('pomodoro-root')
-  )
-}
-
-export function startPomodoro (uuid: string) {
-  if (window?.currentPomodoro?.uuid !== uuid && window?.currentPomodoro?.state?.paused === false) return logseq.UI.showMsg('Another block is running pomodoro timer, please finish it first', 'error')
-  logseq.App.registerUIItem('toolbar', {
-    key: 'logseq-plugin-agenda-pomodoro',
-    template: genToolbarPomodoro(uuid, '--:--', 0),
-  })
-  setTimeout(() => {
-    renderPomodoroApp(uuid)
-    logseq.showMainUI()
-  }, 0)
 }
