+++ src/register/logseq/blockContextMenuItem.ts
@@ -1,0 +1,9 @@
+import { startPomodoro } from '@/register/pomodoro'
+
+
+
+const initializeBlockContextMenuItem = () => {
+  
+}
+
+export default initializeBlockContextMenuItem
